+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 - Standard metrics incoming requests per second
   ([#758](https://github.com/census-instrumentation/opencensus-python/pull/758))
+- Standard metrics incoming requests average execution rate
+  ([#758](https://github.com/census-instrumentation/opencensus-python/pull/758))
 
 ## 0.7.0
 Released 2019-07-31